Move log_event to separate module
